Changed function self-execute + clear timeout
Changed the timer variable to a key in the internal object so that the 'clearTimeout' function will clear the timeout on resize and not just delay a large number of calls to the specified function. Also the changed function is now a self-executing function as it was not being called but simply returning a function that you could then choose to call.